just before persistent add
