+++ backend/beetup-backend.cabal
@@ -13,7 +13,6 @@ common warnings
 executable beetup-backend
     import:             warnings
     main-is:            Main.hs
-    -- other-modules:
     default-extensions: OverloadedStrings
                       , ViewPatterns
                       , DuplicateRecordFields
@@ -28,5 +27,5 @@ executable beetup-backend
                       , HDBC-sqlite3
                       , password
                       , uuid
-    hs-source-dirs:     app
+    hs-source-dirs:     src
     default-language:   Haskell2010
