+++ backend/src/Domain/User.hs
@@ -1,0 +1,10 @@
+module Domain.User where
+
+import Data.UUID (UUID)
+
+data User = User 
+  { id :: UUID
+  , username :: String
+  , name :: String
+  , password :: String
+  }
